requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
psycopg2-binary
sqlalchemy
pydantic[email]==2.4.2
xgboost==2.0.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
python-dotenv
python-multipart